// lib.auth-engine/lib/enums.ts
enum AUTH_STEP {
	DISPATCH = 'DISPATCH',
	SECRET = 'SECRET',
	GENERATE_TOTP = 'GENERATE_TOTP',
	TOTP = 'TOTP',
	GENERATE_CHALLENGE = 'GENERATE_CHALLENGE',
	CHALLENGE_RESPONSE = 'CHALLENGE_RESPONSE',
	RECAPTCHA = 'RECAPTCHA',
	AUTHENTICATED = 'AUTHENTICATED',
	ERROR = 'ERROR'
}

export { AUTH_STEP };